moved auth requests to actions folder
